requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bedrock-agentcore==0.1.7
boto3==1.40.49
botocore==1.40.49
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
docstring-parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx-sse==0.4.2
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
iniconfig==2.1.0
jmespath==1.0.1
jsonschema-specifications==2025.9.1
jsonschema==4.25.1
mcp==1.16.0
opentelemetry-api==1.37.0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-instrumentation==0.58b0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pydantic-core==2.41.1
pydantic-settings==2.11.0
pydantic==2.12.0
pluggy==1.6.0
psutil==7.1.0
pygments==2.19.2
pytest-asyncio==0.26.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
strands-agents==1.11.0
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
wrapt==1.17.3
zipp==3.23.0
